Simplify multithreading allocation in sieve methods
Made changes to the section of code that handles the allocation of threads in different sieve methods for better readability and efficiency. No longer checks if the Sieve of Atkin is used after assigning multiple threads. Instead, the program immediately assigns only one thread if this method is chosen in a more streamlined monothread mode, providing a more streamlined workflow.