Update thread count in monothread tests
Corrected the number of threads in multiple "Test For" blocks for monothread execution from 8 to 1. This change accurately reflects the definition of monothreading, which involves the use of a single thread in a task. This should help prevent potential errors or conflicts resulting from having a higher number of threads indicated for monothread tests.