Add total computation size and improve output formatting
Three changes were made in this commit. Firstly, a variable 'totalSize' was added to count the size of all the computed prime vectors in the multi-threaded model. This will provide insight into how many primes were processed. Additionally, the console output formatting for 'total time elapsed' in the multi-threading model was enhanced. Lastly, the ending console output in the mono-threaded model was revised to improve legibility.