Update GitHub workflows and add test jobs
Renamed the Workflow file to cmake-build-debug.yml to better describe its purpose. Separated the build and test processes for the debug and the release configurations, hence created a new cmake-build-release.yml file. The tests are now run as part of a separate job, to increase efficiency and readability. Added more test cases to cover more methods in both multithread and monothread versions to ensure the overall quality of the code.