"Change logging methods and fix typo in error handling"
In this commit, the logging method was changed from std::clog to std::cout in multiThread.cpp to increase the consistency in the use of output streams across the application. This contributes to more predictable behavior and easier debugging.

In argInput.cpp, a typo in the error handling directive was fixed. The correct directive, NDEBUG, was employed replacing the previous typo "NDEBBUG". This ensures that error handling works as expected in debug mode.