"Moved multithread mode error checks to argInput.cpp
The conditional error checks for 'Sieve of Atkin' and 'Wheel Sieve' trying to use the multithread mode in multiThread.cpp were not the right place for the checks. They have been moved to argInput.cpp, where command line argument checks are handled. This change streamlines the error checking process, attaching checks to the source of potential errors directly."