Refactor argInput and printHelp functions
Refactor the `argInput` function in `argInput.cpp` to handle additional exceptions and provide a link for issue submission. Also, throw an exception for an invalid number of threads.

Refactor the `printHelp` function in `main.h` to provide additional information about running the program in multiThread mode and the unsupported algorithms. Additionally, specify the purpose of the `n` and `outputFile` arguments.

Ignore changes to imports and requires.