Refactor monoThread and multiThread functions and file writing logic
The modifications include refactoring the `monoThread` and `multiThread` functions in the `monoThread.h` and `multiThread.h` files respectively. Additionally, the logic for writing to file in both modes has been modified. These changes aim to improve the efficiency and readability of the code.

Time elapsed for sieving and writing to file is also measured using timer functions.